refactor(cli): extract shared getCliLaunch into cli-entry.ts
Both getDaemonLaunch and getAgentsInvocation hand-built the
node-vs-standalone relaunch command independently (and getAgentsInvocation
used a weaker extension-only check than getDaemonLaunch's shebang-aware
isNodeScriptEntry). Pull the binary-resolution helpers — getAgentsBinPath,
isNodeScriptEntry, getCliLaunch — into a leaf module cli-entry.ts so the
secrets broker can reuse them without the daemon.ts <-> secrets/agent.ts
import cycle. Both daemon builders now delegate to getCliLaunch.
